Refactor output title to OutputTitle component
